// src/views/account/interface.ts
/**
 **********************************************************************
 * This source code is shared for the sole purpose of demo. Do not copy!
 **********************************************************************
 **/

export interface IFormValues {
  name: string;
  website: string;
  phone: string;
  industries: string;
  tags: string[];
  locations: string[];
  timezone: string;
  isLeadEnabled: boolean;
  leadEmails: string[];
}